Ändringar efter fysiskt möte.

+++ src/test/java/org/gradle/DiscountPercentMock.java
@@ -2,6 +2,8 @@ package org.gradle;
 
 import java.util.Date;
 
+import org.gradle.discounts.Discount;
+
 
 public class DiscountPercentMock extends Discount {
 
@@ -24,10 +26,10 @@ public class DiscountPercentMock extends Discount {
 	}
 
 	@Override
-	public int apply(OrderLine orderLine) {
+	public long apply(OrderLine orderLine) {
 		//Compare Discount.productID with orderLine.productID
 		//If true return reduced price otherwise return normal price
-		int price = orderLine.getProductPrice()*orderLine.getProductQuantity();
+		long price = orderLine.getProductPrice()*orderLine.getProductQuantity();
 		if (itemIDs.contains(orderLine.getProductID()))
 			return (int)Math.ceil((float)(price*(100-percent))/100); //Discount
 		else
